+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,7 +1,8 @@
-package nextstep.mvc.controller.asis;
+package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.controller.Controller;
 
 import java.util.Objects;
