+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,7 +2,7 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.controller.Controller;
 
 public class RegisterViewController implements Controller {
