+++ mvc/src/test/java/nextstep/mvc/controller/tobe/HandlerExecutionTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.controller.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 import samples.TestController;
